Make changes to code when getting information for series
Although the frontend works fine when a user tries to view information
on movies and for some series, there happens to be a problem when trying
to get information about series -- in particular when trying to view
series that have lot of seasons and episodes. Due to the performance
bottlenecks, there has been a separation of concerns.

With this commit, the following changes has been made:

* Extract the number of seaonns a movie has and then construct a list
  that houses the url to fetch the information for each series

* Reconstruct the seriesSeasonsSchema in the /app/utils/definitions
  directory to match the information received from the backend